% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict_RECKMON}
\alias{predict_RECKMON}
\title{Predict from a Fitted RECKMON Model}
\usage{
predict_RECKMON(newx, model_fit)
}
\arguments{
\item{newx}{A numeric matrix containing the new data for which to generate
predictions. Each row is an observation and columns must match the
original training data.}

\item{object}{A fitted model object of class `"reckmon"`, as returned by
`main_RECKMON()`.}
}
\value{
A numeric vector of predicted values.
}
\description{
This function generates predictions for new data using a model object
previously fitted by `main_RECKMON`.
}
